Spike out symmetric crypto, archive builder, and CLI
This commit contains a rough end-to-end spike of archive building
functionality, including support for loading symmetric keys
(AES-256-GCM and ChaCha20Poly1305) into keyrings, using those keys to
encrypt encrypt files with a STREAM-based construction, and a CLI which
provides a basic `tar`-like interface.

It's not yet actually useful (e.g. right now it produces empty files
because it doesn't correctly handle emptying the remaining data in the
buffer for the last block), and huge portions of functionality are
missing (all metadata is presently empty), but should provide the basis
for more incremental work.